Compiler: extract lexer component from parser
