+++ Cargo.toml
@@ -5,8 +5,10 @@ members = [
   "compiler/hash-ast",
   "compiler/hash-hir",
   "compiler/hash-interactive",
+  "compiler/hash-lexer",
   "compiler/hash-parser",
   "compiler/hash-reporting",
+  "compiler/hash-token",
   "compiler/hash-typecheck",
   "compiler/hash-utils",
   "compiler/hash-utils/testing-macros",
