+++ compiler/hash-parser/src/lib.rs
@@ -1,11 +1,8 @@
-//! Parser library lib file.
+//! Hash Compiler Parser crate.
 //!
 //! All rights reserved 2022 (c) The Hash Language authors
 #![feature(cell_update)]
 
 pub mod error;
 pub mod gen;
-pub mod lexer;
 pub mod parser;
-pub mod token;
-mod utils;
