+++ compiler/hash-token/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "hash-token"
+version = "0.1.0"
+authors = ["The Hash Language authors"]
+edition = "2021"
+
+[dependencies]
+thiserror = "1.0"
+derive_more = "0.99"
+strum = "0.21"
+strum_macros = "0.21"
+
+hash-ast = {path = "../hash-ast" }
+hash-alloc = {path = "../hash-alloc" }
+hash-source = {path = "../hash-source" }
